app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
argon2-cffi==23.1.0
requests==2.31.0
beautifulsoup4==4.12.2
bleach==6.0.0
markdown==3.5.1
python-dotenv==1.0.0
validators==0.22.0
Pillow==10.0.1
